requirements.txt
Pillow==9.2.0
tqdm==4.64.1

pytest~=7.1.3
PyInputPlus~=0.2.12
requests~=2.28.1